fix tamir suggestions
