+++ quix-frontend/client/src/lib/runner/services/autocomplete/autocomplete-service.ts
@@ -1,23 +1,25 @@
 import { CodeEditorInstance } from '../../../code-editor';
 import { ICompleterItem as AceCompletion } from '../../../code-editor/services/code-editor-completer';
-import { findRelevantPartOfPrefix ,SqlAutocompleter } from "../../../sql-autocomplete/adapter/sql-autocomplete-adapter";
+import { SqlAutocompleter } from "../../../sql-autocomplete/adapter/sql-autocomplete-adapter";
+import { 
+  AddHighlightAndScoreAfterDotObject, 
+  filterAndAddHighlightAndForKeyWord,
+  AddHighlightAndScoreCollumSearch,
+  AddHighlightAndScoreInObjectSearch
+} from "./highLightAndScore";
 // import { BiSqlWebWorkerMngr } from '../../../language-parsers/sql-parser';
 // import { initSqlWorker } from '../workers/sql-parser-worker';
 import {
-  createMatchMask,
-  findAllIndexOf,
   getKeywordsCompletions,
   getQueryAndCursorPositionFromEditor,
 } from './autocomplete-utils';
-import { IDbInfoConfig , DbInfoService } from '../../../sql-autocomplete/db-info';
+import { IDbInfoConfig, DbInfoService } from '../../../sql-autocomplete/db-info';
 import {
   ContextType,
   evaluateContextFromPosition,
   QueryContext,
 } from '../../../sql-autocomplete/sql-context-evaluator';
-// import { SqlAutocompleter } from '../../../sql-autocomplete/adapter/sql-autocomplete-adapter';
 import { IEditSession } from 'brace';
-// import { table } from 'console';
 import { setupOldCompleter } from './old-autocomplete-service';
 
 /* tslint:disable:no-shadowed-variable */
@@ -48,15 +50,21 @@ export async function setupCompleters(
       query,
       position
     );
+    console.log("1.0")
+
     let contextCompletions: Promise<AceCompletion[]> = sqlAutocompleter.getCompletionItemsFromQueryContext(
       queryContext
     );
-    const filteredCompletions : object[] =  (await contextCompletions).filter(obj => obj.value.includes(prefix));
-    
-    if( filteredCompletions.length === 0 ) {
+
+    console.log("2.0")
+
+
+    const filteredCompletions: object[] = (await contextCompletions).filter(obj => obj.value.includes(prefix));
+
+    if (filteredCompletions.length === 0) {
       contextCompletions = sqlAutocompleter.getAllCompletionItemsFromQueryContextCollumn(
-            queryContext
-          );
+        queryContext
+      );
     }
 
     const [keywords, completions] = await Promise.all([
@@ -68,31 +76,34 @@ export async function setupCompleters(
       queryContext.contextType === ContextType.Undefined
         ? keywords
         : completions;
-
+        
     if (prefix) {
       const lowerCasedPrefix = prefix.trim().toLowerCase();
-      if(filteredCompletions.length === 0) {
-        if(prefix.endsWith('.'))  {
-           all = AddHighlightAndScoreAfterDotObject(all,[0],"");
+      if (filteredCompletions.length === 0) {
+        if (prefix.endsWith('.')) {
+          all = AddHighlightAndScoreAfterDotObject(all, [0], "");
         }
-        else  {
-          all = AddHighlightAndScoreInObjectSearch(all ,queryContext , prefix );
+        else {
+          queryContext.contextType === ContextType.Undefined
+        ? all = filterAndAddHighlightAndForKeyWord(all , lowerCasedPrefix)
+        : all = AddHighlightAndScoreInObjectSearch(all, queryContext, lowerCasedPrefix);
         }
       }
-      else  {
+      else {
         all = all.filter(obj => obj.value.includes(prefix));
-        all = AddHighlightAndScoreCollumSearch(all , lowerCasedPrefix)
-    }
+        all = AddHighlightAndScoreCollumSearch(all, lowerCasedPrefix)
+      }
     }
 
     all.forEach(obj => {
-      if(obj.caption?.length>60)  {
-        obj.caption=obj.caption.substring(0,57) + "..."
+      if (obj.caption?.length > 60) {
+        obj.caption = obj.caption.substring(0, 57) + "..."
       }
-      if(!obj.caption && obj.value.length>60)  {
-        obj.caption=obj.value.substring(0,57) + "..."
-      }        
+      if (!obj.caption && obj.value.length > 60) {
+        obj.caption = obj.value.substring(0, 57) + "..."
+      }
     });
+    console.log("all1:" , all)
     return all.sort((a, b) => a.value.localeCompare(b.value));
   };
 
@@ -101,54 +112,4 @@ export async function setupCompleters(
   });
 
   // editorInstance.addOnDemandCompleter(/[\s]+/, completerFn as any);
-}
-function AddHighlightAndScoreAfterDotObject(all: AceCompletion[], indexes: number[], lowerCasedPrefix: string): AceCompletion[] {
-  all = all.reduce((resultArr: AceCompletion[], completionItem) => {
-    completionItem.matchMask = createMatchMask(
-      indexes,
-      100
-    );
-    completionItem.score = 10000 - 0;
-    resultArr.push(completionItem);
-  return resultArr;
-}, []);
-return all;
-}
-
-function AddHighlightAndScoreInObjectSearch(all: AceCompletion[], queryContext: any, prefix: string): AceCompletion[] {
-  all.reduce((resultArr: AceCompletion[], completionItem) => {
-    const relevantPartOfPrefix = findRelevantPartOfPrefix(queryContext.tables , prefix.split('.')).slice(0, -1); //if same problem for both change in function itself
-    const lastDotIndex = relevantPartOfPrefix.lastIndexOf('.');
-    const startOfSearch = lastDotIndex !== -1 ? relevantPartOfPrefix.slice(0, lastDotIndex + 1) : relevantPartOfPrefix;
-    const searchPart = relevantPartOfPrefix.replace(startOfSearch,'');
-    const indexes = findAllIndexOf(completionItem.caption , searchPart);
-    if (indexes.length > 0) {
-      completionItem.matchMask = createMatchMask(
-        indexes,
-        searchPart.length
-      );
-      completionItem.score = 10000 - indexes[0];
-      resultArr.push(completionItem);
-    }
-    return resultArr;
-  }, []);
-return all;
-}
-
-function AddHighlightAndScoreCollumSearch(all: AceCompletion[] , lowerCasedPrefix : string): AceCompletion[] {
-  all.reduce((resultArr: AceCompletion[], completionItem) => {
-    const indexes = findAllIndexOf(completionItem.value, lowerCasedPrefix);
-
-    if (indexes.length > 0) {
-      completionItem.matchMask = createMatchMask(
-        indexes,
-        lowerCasedPrefix.length
-      );
-      completionItem.score = 10000 - indexes[0];
-      resultArr.push(completionItem);
-    }
-
-    return resultArr;
-  }, []);
-return all;
 }
